requirements.txt
amqp==2.3.2
billiard==3.5.0.4
celery==4.2.0
Django==1.11.4
django-celery-beat==1.1.1
django-ckeditor==5.5.0
django-cors-headers==2.4.0
django-js-asset==1.1.0
django-templated-mail==1.1.1
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
djoser==1.2.0
kombu==4.2.1
mysqlclient==1.3.13
Pillow==5.1.0
PyJWT==1.6.4
pytz==2018.4
redis==2.10.6
vine==1.1.4
